Move hooks into hookstate.
