+++ overlord/overlord.go
@@ -41,7 +41,6 @@ import (
 	"github.com/snapcore/snapd/overlord/ifacestate"
 	"github.com/snapcore/snapd/overlord/patch"
 	"github.com/snapcore/snapd/overlord/snapstate"
-	"github.com/snapcore/snapd/overlord/snapstate/hooks"
 	"github.com/snapcore/snapd/overlord/state"
 	"github.com/snapcore/snapd/store"
 )
@@ -103,7 +102,6 @@ func New() (*Overlord, error) {
 	o.hookMgr = hookMgr
 	o.stateEng.AddManager(o.hookMgr)
 
-	hooks.SetupHooks(o.hookMgr)
 	snapMgr, err := snapstate.Manager(s)
 	if err != nil {
 		return nil, err
